Add --changed-only flag to pull command
git-aware pull filtering to only apply translations for keys changed in the current branch.
Useful when working in a feature branch and translation changes have been done in UI/API
for the translations you are working on locally. Of if you do some adjustments after the
GH action have been run, and want to get them locally to commit to branch.